[refact] recruitment 관련 validation 분리 #97
